Fixed some existing bugs. fixed Component_HashImpl to remove warnings

+++ ProjectBane/Bane-Core/Engine/Core/Common.h
@@ -22,7 +22,7 @@ typedef unsigned char byte;
 
 #define ForceInline __forceinline
 	
-ForceInline uint64 GetDJB264BitHash(const char* Ptr, uint Length) 
+ForceInline uint64 GetDJB264BitHash(const char* Ptr, size_t Length) 
 {
 	uint64 Result = 5381;
